Fixes from darwin that they "forgot" to submit upstream.

+++ source/dcc.c
@@ -1,4 +1,4 @@
-/* $EPIC: dcc.c,v 1.115 2005/03/19 03:55:55 jnelson Exp $ */
+/* $EPIC: dcc.c,v 1.116 2005/03/20 03:20:51 jnelson Exp $ */
 /*
  * dcc.c: Things dealing client to client connections. 
  *
@@ -1192,7 +1192,7 @@ const	char 		*text_display, 	/* What to tell the user we sent */
 
 	if (dcc->flags & DCC_QUOTED)
 	{
-		int	len = strlen(text);
+		size_t	len = strlen(text);
 		text = dequote_it(text, &len);
 		writeval = write(dcc->socket, text, len);
 		new_free(&text);
